refactor: Rename `…plugins/aws/package/compile/events/alexaSmartHome.js`
