+++ lib/plugins/aws/package/compile/events/alexa-smart-home.js
