+++ lib/plugins/index.js
@@ -43,7 +43,7 @@ module.exports = [
   require('./aws/package/compile/events/msk/index.js'),
   require('./aws/package/compile/events/alb/index.js'),
   require('./aws/package/compile/events/alexa-skill.js'),
-  require('./aws/package/compile/events/alexaSmartHome.js'),
+  require('./aws/package/compile/events/alexa-smart-home.js'),
   require('./aws/package/compile/events/iot.js'),
   require('./aws/package/compile/events/iotFleetProvisioning.js'),
   require('./aws/package/compile/events/cloudWatchEvent.js'),
